+++ clients/client-core/src/config/mod.rs
@@ -345,7 +345,7 @@ pub struct Client<T> {
     nym_root_directory: PathBuf,
 
     #[serde(skip)]
-    super_struct: PhantomData<*const T>,
+    super_struct: PhantomData<T>,
 }
 
 impl<T: NymConfig> Default for Client<T> {